Refactor to have a separate file to implement the algorithm
The algorithm code was put in a separate Python file, because this
probably is the way that most people work. Maybe they already have a
Python file with their algorithm code.
Also, having a separate file for the algorithm means we can test the
algorithm without using the Docker container. This is much more
convenient than developing in a container.